+++ src/markdown_checker/utils/__init__.py
@@ -2,14 +2,20 @@ from markdown_checker.utils.extract_links import get_links_from_md_file
 from markdown_checker.utils.extract_links import MarkdownLinks
 from markdown_checker.utils.github_env import get_github_repo_blob_url
 from markdown_checker.utils.list_files import get_files_paths_list
+from markdown_checker.utils.pyproject_config import find_pyproject
+from markdown_checker.utils.pyproject_config import load_config_section
+from markdown_checker.utils.pyproject_config import resolve_default_map
 from markdown_checker.utils.spinner import Spinner
 from markdown_checker.utils.spinner import spinner
 
 __all__ = [
     "MarkdownLinks",
     "Spinner",
+    "find_pyproject",
     "get_files_paths_list",
     "get_github_repo_blob_url",
     "get_links_from_md_file",
+    "load_config_section",
+    "resolve_default_map",
     "spinner",
 ]